Exploratory Data Analysis
These files contains the exploratory data analysis conducted in R on the future learn platform dataset in which three areas were analyzed. The first area is regarding the types of personalities of the platform users. Second area is focused on the reasosn gove by the platform users who left  whereas the third area is foucsed on the employment gained by the platform users.